requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-appconf==1.1.0
django-crispy-forms==2.4
django-select2==8.4.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-decouple==3.8
redis==6.1.0
sqlparse==0.5.3
tzdata==2025.2